Fix tests, remove mocha and use ava
